SDKQE-3573 Ensure cleanup removes empty projects (#104)
We have a number of empty projects that are getting left around and not being cleaned up by the jenkins clean up job that runs every 12 hrs.

We should not fast fail on cluster deletion not working (as in a cluster is stuck in destroy_failed state), and should continue and delete all the empty (cbdino) projects).

The fast fail issue is related with removeAll function, where it is fast failing the request if any error is encountered while deleting the clusters/projects.

Also update the cleanup function to remove all the left over projects too after deleting the clusters.