Use a reconciliation loop when rebalancing.
We previously only attempted to execute a rebalance once, and if there
was an issue that prevented it from completing successfully we would not
fail the dino operation, leading to weird test failures.  We now perform
a sort of reconciliation loop for performing rebalances such that
failures are identified and the operation retried.